Split out a Servlet to purely handle listing
