+++ portfolio/src/main/java/com/google/sps/servlets/CommentServlet.java
@@ -33,73 +33,68 @@ import static com.google.sps.utils.Constants.COMMENT_KEY;
 import static com.google.sps.utils.Constants.COMMENT_MAXNUMBER;
 import static com.google.sps.utils.Constants.COMMENT_TIMESTAMP;
 
-/** Servlet that handles getting and posting comment content. */
+/** Servlet that handles getting comment content. */
 @WebServlet("/comment")
 public final class CommentServlet extends HttpServlet {
 
-  private int maxNumberOfComments;
+//  private int maxNumberOfComments;
+//
+//  @Override
+//  public void init() {
+//    this.maxNumberOfComments = 10; // default value
+//  }
+//
+//  @Override
+//  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+//    // Loads the comment from Datastore
+//    List<Comment> comments = getComments();
+//
+//    // Converts into json form
+//    String json = convertToJsonUsingGson(comments);
+//
+//    // Sends the JSON as the response
+//    sendJsonResponse(response, json);
+//  }
 
-  @Override
-  public void init() {
-    this.maxNumberOfComments = 10; // default value
-  }
-  
-  @Override
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    // Loads the comment from Datastore
-    List<Comment> comments = getComments();
-
-    // Converts into json form
-    String json = convertToJsonUsingGson(comments);
-
-    // Sends the JSON as the response
-    sendJsonResponse(response, json);
-  }
-
-  /** Loads the comment from Datastore */
-  private List<Comment> getComments() {
-    Query query = new Query(COMMENT_KEY).addSort(COMMENT_TIMESTAMP, Query.SortDirection.DESCENDING);
-
-    DatastoreService datastoreService = DatastoreServiceFactory.getDatastoreService();
-    PreparedQuery results = datastoreService.prepare(query);
-
-    List<Comment> comments = new ArrayList<>();
-    for (Entity entity: results.asIterable()) {
-      if (comments.size() > this.maxNumberOfComments) {
-        break;
-      }
-
-      // Loads comments from Datastore
-      Comment comment = Comment.CREATOR.fromEntity(entity);
-      comments.add(comment);
-    }
-
-    return comments;
-  }
-
-  /** Converts the list of Comment objecct into json form */
-  private String convertToJsonUsingGson(List<Comment> comments) {
-    Gson gson = new Gson();
-    String json = gson.toJson(comments);
-    return json;
-  }
-
-  /** Sends the JSON as response */
-  private void sendJsonResponse(HttpServletResponse response, String json) throws IOException {
-    response.setContentType("application/json;");
-    response.getWriter().println(json);
-  }
+//  /** Loads the comment from Datastore */
+//  private List<Comment> getComments() {
+//    Query query = new Query(COMMENT_KEY).addSort(COMMENT_TIMESTAMP, Query.SortDirection.DESCENDING);
+//
+//    DatastoreService datastoreService = DatastoreServiceFactory.getDatastoreService();
+//    PreparedQuery results = datastoreService.prepare(query);
+//
+//    List<Comment> comments = new ArrayList<>();
+//    for (Entity entity: results.asIterable()) {
+//      if (comments.size() > this.maxNumberOfComments) {
+//        break;
+//      }
+//
+//      // Loads comments from Datastore
+//      Comment comment = Comment.CREATOR.fromEntity(entity);
+//      comments.add(comment);
+//    }
+//
+//    return comments;
+//  }
+//
+//  /** Converts the list of Comment objecct into json form */
+//  private String convertToJsonUsingGson(List<Comment> comments) {
+//    Gson gson = new Gson();
+//    String json = gson.toJson(comments);
+//    return json;
+//  }
+//
+//  /** Sends the JSON as response */
+//  private void sendJsonResponse(HttpServletResponse response, String json) throws IOException {
+//    response.setContentType("application/json;");
+//    response.getWriter().println(json);
+//  }
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
     // Gets comments from the form
     String commenter = getParameter(request, /*name=*/COMMENT_COMMENTER, /*defaultValue=*/"");
     String content = getParameter(request, /*name=*/COMMENT_CONTENT, /*defaultValue=*/"No comments");
-
-    String maxNumOfCommentStr = getParameter(request, /*name=*/COMMENT_MAXNUMBER, "10");
-    if (maxNumOfCommentStr != null) {
-      this.maxNumberOfComments = Integer.parseInt(maxNumOfCommentStr);
-    }
     // TODO: validate request parameters
 
     // Stores the comment into the Datastore
